FIX error 500 while transferring to default source

+++ app/code/Magento/InventoryCatalog/Model/ResourceModel/BulkInventoryTransfer.php
@@ -122,11 +122,11 @@ class BulkInventoryTransfer
         $finalQuantity = $orgSourceItemQty + $dstSourceItemQty;
 
         if ($orgSourceItem !== null) {
-            $status = $orgSourceItem[SourceItemInterface::STATUS];
+            $status = (int) $orgSourceItem[SourceItemInterface::STATUS];
         } elseif ($dstSourceItemQty !== null) {
-            $status = $dstSourceItem[SourceItemInterface::STATUS];
+            $status = (int) $dstSourceItem[SourceItemInterface::STATUS];
         } else {
-            $status = SourceItemInterface::STATUS_OUT_OF_STOCK;
+            $status = (int) SourceItemInterface::STATUS_OUT_OF_STOCK;
         }
 
         $updateOperation = [
